[#239] Added Angle and Flip to texture rendering
+ Still need to implement these for ImGUI textures. [#257] [#258]

+++ SuperGameEngine/Engine/Structural/Assets/Texture/SuperTextureAsset.cpp
@@ -132,8 +132,7 @@ void SuperTextureAsset::Draw(int tile, const FatedQuestLibraries::FVector2F& scr
             DrawPredefined(tile, screenLocation, tintColour);
             break;
         default:
-            // TODO: Implment Tint here
-            m_superTexture->Draw(FPoint(screenLocation.GetX(), screenLocation.GetY()), tintColour);
+            m_superTexture->Draw(FPoint(static_cast<int>(screenLocation.GetX()), static_cast<int>(screenLocation.GetY())), tintColour);
         }
     }
 }
